OpenCL Version 0.0.3
Finished being able to hash out the pre-compute and post-compute of the xor_salsa on the host so that we can just focus on dancing with the GPU.  In addition, we can decide how many individual threads we want to upload to a card at a time.  Since this is the first version, I've hard coded this to only work with the 4 way algorithm.